+++ example/test/step/i_do_not_see_text.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_test/flutter_test.dart';
+
+Future<void> iDoNotSeeText(
+  WidgetTester tester,
+  String text,
+) async {
+  expect(find.text(text), findsNothing);
+}