+++ lib/src/step_generator.dart
@@ -36,6 +36,7 @@ String getStepFilename(String stepText) {
 String getStepMethodName(String stepText) {
   final text = stepText
       .replaceAll(parametersRegExp, '')
+      .replaceAll(examplesRegExp, '')
       .replaceAll(charactersAndNumbersRegExp, '')
       .replaceAll(repeatingSpacesRegExp, ' ')
       .trim()