+++ test/after_test.dart
@@ -26,8 +26,8 @@ void main() {
   Future<void> bddTearDown(WidgetTester tester) async {
     await theTestFinishes(tester);
   }
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await testingScenario();
       await theAppIsRunning(tester);
       await bddTearDown(tester);