+++ test/background_test.dart
@@ -26,8 +26,8 @@ void main() {
   Future<void> bddSetUp(WidgetTester tester) async {
     await theServerAlwaysReturnErrors(tester);
   }
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await bddSetUp(tester);
       await testingScenario();
       await theAppIsRunning(tester);