+++ test/feature_test.dart
@@ -35,8 +35,8 @@ void main() {
 import './step/the_app_is_running.dart';
 
 void main() {
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await testingScenario();
       await theAppIsRunning(tester);
     });
@@ -61,8 +61,8 @@ import './step/i_see_text.dart';
 import './step/i_see_icon.dart';
 
 void main() {
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await testingScenario();
       await theAppIsRunning(tester);
       await iSeeText(tester, 'nice param');
@@ -89,18 +89,18 @@ import './step/the_app_is_running.dart';
 import './step/second_testing_scenario.dart';
 
 void main() {
-  group('First testing feature', () {
-    testWidgets('First testing scenario', (tester) async {
+  group(\'\'\'First testing feature\'\'\', () {
+    testWidgets(\'\'\'First testing scenario\'\'\', (tester) async {
       await firstTestingScenario();
       await theAppIsRunning(tester);
     });
   });
-  group('Second testing feature', () {
-    testWidgets('First testing scenario', (tester) async {
+  group(\'\'\'Second testing feature\'\'\', () {
+    testWidgets(\'\'\'First testing scenario\'\'\', (tester) async {
       await firstTestingScenario();
       await theAppIsRunning(tester);
     });
-    testWidgets('Second testing scenario', (tester) async {
+    testWidgets(\'\'\'Second testing scenario\'\'\', (tester) async {
       await secondTestingScenario();
       await theAppIsRunning(tester);
     });