+++ test/setup_test.dart
@@ -32,8 +32,8 @@ void main() {
   Future<void> bddTearDown(WidgetTester tester) async {
     await theTestFinishes(tester);
   }
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await bddSetUp(tester);
       await testingScenario();
       await theAppIsRunning(tester);