+++ test/step_folder_name_test.dart
@@ -21,8 +21,8 @@ import './../../../custom_steps/testing_scenario.dart';
 import './../../../custom_steps/the_app_is_running.dart';
 
 void main() {
-  group('Testing feature', () {
-    testWidgets('Testing scenario', (tester) async {
+  group(\'\'\'Testing feature\'\'\', () {
+    testWidgets(\'\'\'Testing scenario\'\'\', (tester) async {
       await testingScenario();
       await theAppIsRunning(tester);
     });