+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.2.0] - Scenario Outline
+
+* Add scenario outline support
+
+## [1.1.0] - Integration tests
+
+* Add support for integration tests
+
 ## [1.0.3+1] - Custom step folder name hotix
 
 * Hotfix for ".." in custom step names